# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/user": {
            "get": {
                "description": "用户分页",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User 用户"
                ],
                "summary": "用户分页",
                "operationId": "UserIndex",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/resp.UserResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/add": {
            "post": {
                "description": "添加用户相关接口",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User 用户"
                ],
                "summary": "添加用户",
                "operationId": "UserAdd",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.UserAddReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/delete/{id}": {
            "post": {
                "description": "删除用户相关接口",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User 用户"
                ],
                "summary": "删除用户",
                "operationId": "UserDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.IdReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/detail/{id}": {
            "get": {
                "description": "用户详情相关接口",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User 用户"
                ],
                "summary": "用户详情",
                "operationId": "UserDetail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/resp.UserResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/disable": {
            "post": {
                "description": "禁用用户相关接口",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User 用户"
                ],
                "summary": "禁用用户",
                "operationId": "UserDisable",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.IdReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/edit": {
            "post": {
                "description": "编辑用户相关接口",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User 用户"
                ],
                "summary": "编辑用户",
                "operationId": "UserEdit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.UserEditReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/list": {
            "get": {
                "description": "用户列表相关接口",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User 用户"
                ],
                "summary": "用户列表",
                "operationId": "UserList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "查询条件",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.UserQueryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.PageResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/self": {
            "get": {
                "description": "用户权限相关接口",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User 用户"
                ],
                "summary": "用户权限",
                "operationId": "UserSelf",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/resp.UserSelfResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/update": {
            "post": {
                "description": "更新用户相关接口",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User 用户"
                ],
                "summary": "更新用户",
                "operationId": "UserUpdate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.UserUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "req.IdReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "主键ID",
                    "type": "integer"
                }
            }
        },
...
        "req.UserAddReq": {
            "type": "object",
            "required": [
                "password",
                "phone",
                "username"
            ],
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "is_disable": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "is_multipoint": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 6
                },
                "phone": {
                    "type": "string"
                },
                "role_id": {
                    "type": "integer",
                    "minimum": 0
                },
                "tenant_id": {
                    "type": "integer",
                    "minimum": 0
                },
                "username": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 4
                }
            }
        },
        "req.UserEditReq": {
            "type": "object",
            "required": [
                "id",
                "phone",
                "username"
            ],
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "is_disable": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "is_multipoint": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "nickname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role_id": {
                    "type": "integer",
                    "minimum": 0
                },
                "tenant_id": {
                    "type": "integer",
                    "minimum": 0
                },
                "username": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 4
                }
            }
        },
        "req.UserQueryReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "is_disable": {
                    "type": "integer",
                    "default": -1,
                    "enum": [
                        0,
                        1,
                        -1
                    ]
                },
                "phone": {
                    "type": "string"
                },
                "role_id": {
                    "type": "integer",
                    "minimum": 0
                },
                "tenant_id": {
                    "type": "integer",
                    "minimum": 0
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "req.UserUpdateReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "confirm_password": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 6
                }
            }
        },
...
        "resp.UserResp": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "created_at": {
                    "$ref": "#/definitions/types.DateTime"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_disable": {
                    "type": "integer"
                },
                "is_multipoint": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "integer"
                        },
                        "is_admin": {
                            "type": "integer"
                        },
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "tenant": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "integer"
                        },
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "updated_at": {
                    "$ref": "#/definitions/types.DateTime"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "resp.UserSelfResp": {
            "type": "object",
            "properties": {
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user": {
                    "$ref": "#/definitions/resp.UserResp"
                }
            }
        },
        "response.PageResp": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "总数",
                    "type": "integer"
                },
                "lists": {
                    "description": "数据"
                },
                "pageNo": {
                    "description": "页No",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页Size",
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "types.DateTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        }
    }
}